fix(nexus js - three.js): Correct bounding sphere test for raycasting
The bounding sphere test which allows raycasting to exit early is now
working properly thanks to transforming the bounding sphere of
the original (i.e. untransformed) Nexus mesh by the world matrix of
the displayed (i.e. transformed) mesh.

As a consequence, this fix makes raycasting to work also on models
that are originally very small (whose bounding sphere radius < 1).